+++ docs/additional_info/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 4.1.0 (26-Feb-21)
+
+* Added `payment_method` and `dry_run` attributes to `Order`
+
 ## 4.0.0 (02-Feb-21)
 
 * **License type is now BSD 3 Clause**